+++ src/main/java/com/ssafy/ssafsound/domain/lunch/dto/GetScrapResDto.java
@@ -1,0 +1,4 @@
+package com.ssafy.ssafsound.domain.lunch.dto;
+
+public interface GetScrapResDto {
+}
