+++ src/main/java/com/ssafy/ssafsound/domain/lunch/service/LunchScrapService.java
@@ -8,12 +8,9 @@ import com.ssafy.ssafsound.domain.lunch.dto.GetLunchListReqDto;
 import com.ssafy.ssafsound.domain.lunch.exception.LunchErrorInfo;
 import com.ssafy.ssafsound.domain.lunch.exception.LunchException;
 import com.ssafy.ssafsound.domain.lunch.repository.LunchRepository;
-import com.ssafy.ssafsound.domain.lunch.task.domain.ScrapInfoProvider;
-import com.ssafy.ssafsound.domain.lunch.task.domain.ScrapInfoProviderFactory;
-import com.ssafy.ssafsound.domain.lunch.task.domain.WelstoryInfoProvider;
-import com.ssafy.ssafsound.domain.lunch.task.dto.GetFreshmealResDto;
-import com.ssafy.ssafsound.domain.lunch.task.dto.GetScrapReqDto;
-import com.ssafy.ssafsound.domain.lunch.task.dto.GetWelstoryResDto;
+import com.ssafy.ssafsound.domain.lunch.dto.GetFreshmealResDto;
+import com.ssafy.ssafsound.domain.lunch.dto.GetScrapReqDto;
+import com.ssafy.ssafsound.domain.lunch.dto.GetWelstoryResDto;
 import com.ssafy.ssafsound.domain.member.exception.MemberErrorInfo;
 import com.ssafy.ssafsound.domain.meta.domain.Campus;
 import com.ssafy.ssafsound.domain.meta.domain.MetaData;
