+++ src/main/java/com/ssafy/ssafsound/domain/lunch/service/ScrapInfoProvider.java
@@ -1,0 +1,8 @@
+package com.ssafy.ssafsound.domain.lunch.service;
+
+import com.ssafy.ssafsound.domain.lunch.dto.GetScrapReqDto;
+import com.ssafy.ssafsound.domain.lunch.dto.GetScrapResDto;
+
+public interface ScrapInfoProvider {
+    GetScrapResDto scrapLunchInfo(GetScrapReqDto getScrapReqDto);
+}
