--- src/main/java/com/ssafy/ssafsound/domain/lunch/task/dto/GetScrapResDto.java
@@ -1,4 +1,0 @@
-package com.ssafy.ssafsound.domain.lunch.task.dto;
-
-public interface GetScrapResDto {
-}
