+++ src/main/java/com/ssafy/ssafsound/global/config/ScrapConfig.java
@@ -1,7 +1,7 @@
 package com.ssafy.ssafsound.global.config;
 
-import com.ssafy.ssafsound.domain.lunch.task.domain.FreshmealProperties;
-import com.ssafy.ssafsound.domain.lunch.task.domain.WelstoryProperties;
+import com.ssafy.ssafsound.domain.lunch.domain.FreshmealProperties;
+import com.ssafy.ssafsound.domain.lunch.domain.WelstoryProperties;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
